refactor(path): change the function position
- 각 경로 조회 함수 위치 변경
